d/mailman3-web.postinst: Encode the output given by django-admin as soon as it can be non-ascii
Tweak a little the django admin call used to fetch Site.objects.all()[0].name,
as python can't encode automagically its output in a subshell (as it can't with
pipes). Use locale module with getpreferredencoding function to set a default
encoding that should work on the current machine running the postinst script.